type error fix
