+++ components/types/Comment.ts
@@ -1,0 +1,12 @@
+interface Comment {
+    commentId: string;
+    content: string;
+    author: {
+      username: string;
+      nickname: string;
+      profilePictureURL: string;
+    };
+    creationDate: string;
+  }
+
+  export default Comment;
