+++ screens/FeedScreen.tsx
@@ -13,11 +13,12 @@ import ErrorComp from "../components/ErrorComp";
 import { useAuth } from "../authentification/AuthContext";
 import { navigate } from "../navigation/NavigationService";
 import { NativeScrollEvent } from "react-native";
+import Post from "../components/types/Post";
 
 const FeedScreen = () => {
   const { token } = useAuth();
-  const [postsPersonal, setPostsPersonal] = useState([]);
-  const [postsGlobal, setPostsGlobal] = useState([]);
+  const [postsPersonal, setPostsPersonal] = useState<Post[]>([]);
+  const [postsGlobal, setPostsGlobal] = useState<Post[]>([]);
   const [errorText, setErrorText] = useState("");
   const [loading, setLoading] = useState(false);
   const [refreshing, setRefreshing] = useState(false);
@@ -96,6 +97,7 @@ const FeedScreen = () => {
     }
     setRefreshing(true);
     fetchPosts("personal");
+    fetchPosts("global");
   };
 
   const handleScroll = ({
@@ -130,7 +132,7 @@ const FeedScreen = () => {
     }
   };
 
-  const loadCitiesForPosts = async (posts: any) => {
+  const loadCitiesForPosts = async (posts: Post[]) => {
     const updatedPosts = [];
     for (const post of posts) {
       if (post.location && post.location.latitude && post.location.longitude) {
@@ -169,7 +171,6 @@ const FeedScreen = () => {
               date={post.creationDate}
               postContent={post.content}
               city={loadingCities ? "Loading city..." : post.city}
-              token={token}
               initialLikes={post.likes}
               initialLiked={post.liked}
             />
@@ -199,9 +200,9 @@ const FeedScreen = () => {
           date={post.creationDate}
           postContent={post.content}
           city={loadingCities ? "Loading city..." : post.city}
-          token={token}
           initialLikes={post.likes}
           initialLiked={post.liked}
+        
         />
       ))}
       {hasMoreGlobalPosts && (
